Extract cli code
