+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './bundler';
 export * from './runner';
+export * from './cli';
